Added follow and unfollow methods + followersCount
followersCount is a number property of an Event.
It's incremented and decremented each time a user follows an unfollows
the event respectively.
It's a lesser intensive computation than run through all related
model's relations.
But it's also less accurate and might not exactly reflect the real
number of followers of the event.
A good thing would be a job to periodically check and adjust this
number.
see https://github.com/strongloop/loopback/issues/917